+++ src/Server/socket/listen.cpp
@@ -2,7 +2,7 @@
 
 void Server::startListen(std::vector<t_multi_server> &servers, int back_log)
 {
-  for (int i = 0; i < servers.size(); ++i)
+  for (unsigned long i = 0; i < servers.size(); ++i)
   {
     if (listen(servers[i].server_sock, back_log) == -1)
     {
